[#963] updates create-app template
