+++ packages/create-app/templates/default-app/packages/api/package.json.hbs
@@ -81,7 +81,9 @@
     "cors": "^2.8.5",
     "express": "^4.17.1",
     "helmet": "^5.0.1",
-    "module-alias": "^2.2.2"
+    "module-alias": "^2.2.2",
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^4.6.2"
   },
   "lint-staged": {
     "*.{js,ts}": [
