+++ packages/create-app/templates/default-app/packages/api/src/api.ts
@@ -3,131 +3,181 @@
  */
 
 import express from 'express'
-
+import { setConfig, CCFConfig } from '@cloud-carbon-footprint/common'
 import {
-  App,
-  createValidFootprintRequest,
-  createValidRecommendationsRequest,
-  FootprintEstimatesRawRequest,
-  RecommendationsRawRequest,
-  Tags,
-} from '@cloud-carbon-footprint/app'
-
-import {
-  setConfig,
-  CCFConfig,
-  EstimationRequestValidationError,
-  Logger,
-  PartialDataError,
-  RecommendationsRequestValidationError,
-} from '@cloud-carbon-footprint/common'
-
-const apiLogger = new Logger('API')
-
-/**
- * Returns the raw estimates
- *
- * Query params:
- * start - Required, UTC start date in format YYYY-MM-DD
- * end - Required, UTC start date in format YYYY-MM-DD
- */
-const FootprintApiMiddleware = async function (
-  req: express.Request,
-  res: express.Response,
-): Promise<void> {
-  // Set the request time out to 10 minutes to allow the request enough time to complete.
-  req.socket.setTimeout(1000 * 60 * 10)
-  const rawRequest: FootprintEstimatesRawRequest = {
-    startDate: req.query.start?.toString(),
-    endDate: req.query.end?.toString(),
-    ignoreCache: req.query.ignoreCache?.toString(),
-    groupBy: req.query.groupBy?.toString(),
-    limit: req.query.limit?.toString(),
-    skip: req.query.skip?.toString(),
-    cloudProviders: req.query.cloudProviders as string[],
-    accounts: req.query.accounts as string[],
-    services: req.query.services as string[],
-    regions: req.query.regions as string[],
-    tags: req.query.tags as Tags,
-  }
-  apiLogger.info(`Footprint API request started.`)
-  if (!rawRequest.groupBy) {
-    apiLogger.warn('GroupBy parameter not specified, adopting default "day"')
-    rawRequest.groupBy = 'day'
-  }
-  const footprintApp = new App()
-  try {
-    const estimationRequest = createValidFootprintRequest(rawRequest)
-    const estimationResults = await footprintApp.getCostAndEstimates(
-      estimationRequest,
-    )
-    res.json(estimationResults)
-  } catch (e) {
-    apiLogger.error(`Unable to process footprint request.`, e)
-    if (
-      e.constructor.name ===
-      EstimationRequestValidationError.prototype.constructor.name
-    ) {
-      res.status(400).send(e.message)
-    } else if (
-      e.constructor.name === PartialDataError.prototype.constructor.name
-    ) {
-      res.status(416).send(e.message)
-    } else res.status(500).send('Internal Server Error')
-  }
-}
-
-const EmissionsApiMiddleware = async function (
-  req: express.Request,
-  res: express.Response,
-): Promise<void> {
-  apiLogger.info(`Regions emissions factors API request started`)
-  const footprintApp = new App()
-  try {
-    const emissionsResults = await footprintApp.getEmissionsFactors()
-    res.json(emissionsResults)
-  } catch (e) {
-    apiLogger.error(`Unable to process regions emissions factors request.`, e)
-    res.status(500).send('Internal Server Error')
-  }
-}
-
-const RecommendationsApiMiddleware = async function (
-  req: express.Request,
-  res: express.Response,
-): Promise<void> {
-  const rawRequest: RecommendationsRawRequest = {
-    awsRecommendationTarget: req.query.awsRecommendationTarget?.toString(),
-  }
-  apiLogger.info(`Recommendations API request started`)
-  const footprintApp = new App()
-  try {
-    const estimationRequest = createValidRecommendationsRequest(rawRequest)
-    const recommendations = await footprintApp.getRecommendations(
-      estimationRequest,
-    )
-    res.json(recommendations)
-  } catch (e) {
-    apiLogger.error(`Unable to process recommendations request.`, e)
-    if (
-      e.constructor.name ===
-      RecommendationsRequestValidationError.prototype.constructor.name
-    ) {
-      res.status(400).send(e.message)
-    } else {
-      res.status(500).send('Internal Server Error')
-    }
-  }
-}
+  FootprintApiMiddleware,
+  EmissionsApiMiddleware,
+  RecommendationsApiMiddleware,
+} from './middleware'
 
 export const createRouter = (config?: CCFConfig) => {
   setConfig(config)
 
   const router = express.Router()
 
+  /**
+   * @openapi
+   * /api/footprint:
+   *  get:
+   *     tags:
+   *     - Footprint
+   *     summary: Gets calculated energy and carbon estimates for a given date range
+   *     produces:
+   *       - application/json
+   *     parameters:
+   *      - name: start
+   *        in: query
+   *        description: The start date for the footprint; e.g. 2022-10-18
+   *        schema:
+   *          type: string
+   *        required: true
+   *      - name: end
+   *        in: query
+   *        schema:
+   *          type: string
+   *        description: The end date for the footprint
+   *        required: true
+   *      - name: ignoreCache
+   *        in: query
+   *        schema:
+   *          type: boolean
+   *          default: false
+   *        required: false
+   *      - name: groupBy
+   *        in: query
+   *        schema:
+   *          type: string
+   *          default: day
+   *        required: false
+   *      - name: limit
+   *        in: query
+   *        schema:
+   *          type: number
+   *          default: 50000
+   *        description: The maximum number of estimates to return (MongoDB only, ignoreCache=false)
+   *        required: false
+   *      - name: skip
+   *        in: query
+   *        schema:
+   *          type: number
+   *          default: 0
+   *        description: The maximum number of estimates to skip over (MongoDB only, ignoreCache=false)
+   *        required: false
+   *      - name: cloudProviders
+   *        in: query
+   *        schema:
+   *          type: array
+   *          items:
+   *            type: string
+   *        description: List of Cloud Providers to include in estimates (MongoDB only, Filter)
+   *        required: false
+   *      - name: accounts
+   *        in: query
+   *        schema:
+   *          type: array
+   *          items:
+   *            type: string
+   *        description: List of accounts to include in estimates (MongoDB only, Filter)
+   *        required: false
+   *      - name: services
+   *        in: query
+   *        schema:
+   *          type: array
+   *          items:
+   *            type: string
+   *        description: List of services to include in estimates (MongoDB only, Filter)
+   *        required: false
+   *      - name: regions
+   *        in: query
+   *        schema:
+   *          type: array
+   *          items:
+   *            type: string
+   *        description: List of regions to include in estimates (MongoDB only, Filter)
+   *        required: false
+   *      - name: tags
+   *        in: query
+   *        schema:
+   *          type: object
+   *          additionalProperties:
+   *            type: string
+   *        description: List of resource tags to include in estimates (MongoDB only, Filter)
+   *        required: false
+   *     responses:
+   *       200:
+   *         description: Success
+   *         content:
+   *          application/json:
+   *            schema:
+   *                type: array
+   *                items:
+   *                  $ref: '#/components/schemas/FootprintResponse'
+   *       400:
+   *         description: Bad request
+   *       416:
+   *         description: Partial Data Error
+   *       500:
+   *         description: Internal Server Error
+   */
   router.get('/footprint', FootprintApiMiddleware)
+
+  /**
+   * @openapi
+   * /api/regions/emissions-factors:
+   *  get:
+   *     tags:
+   *     - Emissions Factors
+   *     description: Gets the carbon intensity (co2e/kWh) of all cloud provider regions
+   *     responses:
+   *       200:
+   *         description: Success
+   *         content:
+   *           application/json:
+   *            schema:
+   *              type: array
+   *              items:
+   *                $ref: '#/components/schemas/EmissionResponse'
+   */
   router.get('/regions/emissions-factors', EmissionsApiMiddleware)
+
+  /**
+   * @openapi
+   * /api/recommendations:
+   *  get:
+   *     tags:
+   *     - Recommendations
+   *     description: Gets recommendations from cloud providers and their estimated carbon and energy impact
+   *     parameters:
+   *      - name: awsRecommendationTarget
+   *        in: query
+   *        description: Defines whether targeted AWS recommendations should be within the same family
+   *        schema:
+   *          type: string
+   *          enum: [SAME_INSTANCE_FAMILY, CROSS_INSTANCE_FAMILY]
+   *        required: true
+   *     responses:
+   *       200:
+   *         description: Success
+   *         content:
+   *           application/json:
+   *            schema:
+   *              type: array
+   *              items:
+   *                $ref: '#/components/schemas/RecommendationsResponse'
+   */
   router.get('/recommendations', RecommendationsApiMiddleware)
+
+  /**
+   * @openapi
+   * /api/healthz:
+   *  get:
+   *     tags:
+   *     - Healthcheck
+   *     description: Responds if the app is up and running
+   *     responses:
+   *       200:
+   *         description: Responds "OK" if app is up and running
+   */
   router.get('/healthz', (req: express.Request, res: express.Response) => {
     res.status(200).send('OK')
   })
