+++ packages/create-app/templates/default-app/packages/api/src/server.ts
@@ -11,9 +11,10 @@ import helmet from 'helmet'
 import cors, { CorsOptions } from 'cors'
 
 import { createRouter } from './api'
-import auth from './auth'
 import { Logger, configLoader } from '@cloud-carbon-footprint/common'
-import MongoDbCacheManager from '@cloud-carbon-footprint/app/src/MongoDbCacheManager'
+import { MongoDbCacheManager } from '@cloud-carbon-footprint/app'
+import swaggerDocs from './utils/swagger'
+import auth from './utils/auth'
 
 const port = process.env.PORT || 4000
 const httpApp = express()
@@ -47,11 +48,12 @@ if (process.env.ENABLE_CORS) {
 
 httpApp.use('/api', createRouter())
 
-httpApp.listen(port, () =>
+httpApp.listen(port, () => {
   serverLogger.info(
     `Cloud Carbon Footprint Server listening at http://localhost:${port}`,
-  ),
-)
+  )
+  swaggerDocs(httpApp, Number(port))
+})
 
 // Instructions for graceful shutdown
 process.on('SIGINT', async () => {
